pyproject.toml
[project]
name = "train-punctuality-service"
version = "0.1.0"
description = "Indian Railways train punctuality prediction service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "apscheduler>=3.10.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "prophet>=1.1.5",
    "fake-useragent>=1.5.0",
    "python-dotenv>=1.0.0",
    "pyarrow>=15.0.0",
    "libsql-experimental>=0.0.68",
    "playwright>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
